+++ pyTRT/methods/ILS.py
@@ -5,8 +5,10 @@ import math
 
 import numpy as np
 
+from GHEtool.VariableClasses.GroundData._GroundData import _GroundData
 from pyTRT.utils import TRTData
 from pyTRT.methods.baseclass import _Method
+from typing import Union
 
 
 class ILS(_Method):
@@ -14,7 +16,8 @@ class ILS(_Method):
     This class contains the infinite line source (ILS) method for the analysis of the TRT measurements.
     """
 
-    def __init__(self, data: TRTData, borehole_length: float, borehole_radius: float, volumetric_heat_capacity: float):
+    def __init__(self, data: TRTData, borehole_length: float, borehole_radius: float,
+                 volumetric_heat_capacity: Union[float, _GroundData]):
         """
         Initialises the infinite line source (ILS) method for the analysis of the TRT measurements.
         This method is based on the work of (Gehlin, S., 2002) [#Gehlin]_.
@@ -27,7 +30,7 @@ class ILS(_Method):
             Length of the borehole heat exchanger [m]
         borehole_radius : float
             Radius of the borehole [m]
-        volumetric_heat_capacity : float
+        volumetric_heat_capacity : float | _GroundData
             Volumetric heat capacity [J/(m³K)]
 
         References
@@ -35,6 +38,10 @@ class ILS(_Method):
         .. [#gehlin2002] Gehlin, S. (2002). *Thermal Response Test: Method, Development and Evaluation* (Ph.D. dissertation).
            Department of Environmental Engineering, Luleå University of Technology, Sweden.
         """
+
+        # GHEtool object so convert to volumetric heat capacity
+        if isinstance(volumetric_heat_capacity, _GroundData):
+            volumetric_heat_capacity = volumetric_heat_capacity.volumetric_heat_capacity(borehole_length, 1)
 
         a, b = np.polyfit(data.log_time_array, data.temperature_array, 1)
 
@@ -47,3 +54,7 @@ class ILS(_Method):
 
         # initiate baseclass
         super().__init__(Rb, ks)
+
+        # save correlation parameters
+        self._a = a
+        self._b = b