+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.0.3] - 2025-08-18
+
+### Added
+
+- Option to use GHEtool ground object in ILS.
+- Save a and b from correlation in data
+
+### Changed
+
+- Update documentation with installation instructions.
+
 ## [0.0.2] - 2025-07-22
 
 ### Changed